+++ ChangeLog.md
@@ -114,6 +114,8 @@
 
   * Add new python grading framework for exampleCourse (Nathan Walters).
 
+  * Add CSV download for instance questions (Matt West).
+
   * Split installing documentation into separate method sections (Matt West).
 
   * Remove unused dead code (`/lib/db.js`, `question-servers/shortAnswer.js`,